feat: 와인노트상세 화면 추가 작업
