+++ package.json
@@ -30,6 +30,7 @@
     "react-router-dom": "^5.3.0",
     "react-scripts": "^4.0.3",
     "react-select": "^5.1.0",
+    "react-star-ratings": "^2.3.0",
     "recoil": "^0.4.1"
   },
   "devDependencies": {
